// shipeasy/src/app/models/system-type.ts
export interface SystemType {
    systemtypeId: string;
    createdBy: string;
    createdOn: string;
    module: string;
    orgId: string;
    status: boolean;
    tenantId: string;
    typeActive: boolean;
    typeCategory: string;
    typeDescription: string;
    typeName: string;
    typeParentType: string;
    typeRef: string;
    typeRefId: string;
    updatedBy: string;
    updatedOn: string;
    processType:string;
    item_text:string;
    item_id:string;
  }
  